completely restructured how the contact form linking worked
fixes SPOT-571
refs SDEV-115

now, instead of setting session variables with the spot info,
the spot_id is part of the url and then the other information
is parsed out of a json call. contact.py was also redone so
that it only queries the server for the json data once per
view. it's more effecient and consistent